Updated MergeByTimeProcessor to new API
